+++ internlm/core/scheduler/no_pipeline_scheduler.py
@@ -202,8 +202,10 @@ class NonPipelineScheduler(BaseScheduler):
             if return_output_label:
                 outputs.append(_output)
                 labels.append(_label)
+
         if not return_output_label:
             outputs, labels = None, None
+
         # Compatible for non-moe
         if hasattr(gpc.config.model, "num_experts"):
             return outputs, labels, loss, moe_loss
